refactor : Presenter의 UI 로직 제거
