+++ app/src/main/java/woowacourse/movie/presenter/seat/SeatPresenter.kt
@@ -2,9 +2,7 @@ package woowacourse.movie.presenter.seat
 
 import woowacourse.movie.domain.model.seat.Seat
 import woowacourse.movie.domain.model.seat.Seats
-import woowacourse.movie.view.seat.model.coord.Column
 import woowacourse.movie.view.seat.model.coord.Coordination
-import woowacourse.movie.view.seat.model.coord.Row
 
 class SeatPresenter(
     private val view: SeatContract.View,
@@ -18,7 +16,7 @@ class SeatPresenter(
         val changed = seats.toggleSeat(newSeat, limit)
 
         if (changed) {
-            view.showSeat(seatToCoordination())
+            view.showSeat(seats.item)
             view.showPrice(seats.bookingPrice())
             updateConfirmButtonState(limit)
         } else {
@@ -31,29 +29,11 @@ class SeatPresenter(
             view.showToast(limit)
             return
         }
-        view.moveToBookingComplete(seatToLabel(), seats.bookingPrice())
+        view.moveToBookingComplete(seats.item, seats.bookingPrice())
     }
 
     private fun updateConfirmButtonState(peopleCount: Int) {
         val isEnabled = seats.item.size == peopleCount
         view.setConfirmButtonEnabled(isEnabled)
-    }
-
-    private fun seatToCoordination(): List<Coordination> {
-        return seats
-            .item
-            .map {
-                Coordination(Column(it.x), Row(it.y))
-            }
-    }
-
-    private fun seatToLabel(): String {
-        return seats
-            .item
-            .joinToString {
-                val rowLetter = ('A' + it.x - 1)
-                val columnNumber = it.y
-                "$rowLetter$columnNumber"
-            }
     }
 }
